Introduce abstract Cellpose class
