+++ src/main/java/org/mastodon/mamut/detection/Cellpose3.java
@@ -2,26 +2,57 @@ package org.mastodon.mamut.detection;
 
 import java.io.IOException;
 
-public class Cellpose3 extends Segmentation3D
+public class Cellpose3 extends Cellpose
 {
 	private final MODEL_TYPE modelType;
-
 	private double anisotropy = 1;
-
-	private double cellprobThreshold = 0;
-
-	private double flowThreshold = 0;
-
-	private boolean is3D = true;
-
-	public final static double DEFAULT_CELLPROB_THRESHOLD = 3d;
-
-	public final static double DEFAULT_FLOW_THRESHOLD = 0.4d;
 
 	public Cellpose3( final MODEL_TYPE modelType ) throws IOException
 	{
 		super();
 		this.modelType = modelType;
+	}
+
+	public double getAnisotropy()
+	{
+		return anisotropy;
+	}
+
+	public void setAnisotropy( final double anisotropy )
+	{
+		this.anisotropy = anisotropy;
+	}
+
+	private String anisotropyParam()
+	{
+		return is3D() ? String.valueOf( anisotropy ) : "1.0";
+	}
+
+	@Override
+	protected String getLoadModelCommand()
+	{
+		if ( modelType.hasSizeModel() )
+			return "model = models.Cellpose(model_type=\"" + modelType.getModelName() + "\", gpu=True)" + "\n";
+		else
+			return "model = models.CellposeModel(model_type=\"" + modelType.getModelName() + "\", gpu=True)" + "\n";
+	}
+
+	@Override
+	protected String getEvaluateModelCommand()
+	{
+		String diams = modelType.hasSizeModel() ? ", diams" : "";
+		return "segmentation, flows, styles" + diams + " = model.eval("
+				+ "image_ndarray, "
+				+ "diameter=None, "
+				+ "channels=[0, 0], "
+				+ "do_3D=" + is3DParam() + ", "
+				+ "anisotropy=" + anisotropyParam() + ", "
+				+ "z_axis=0, "
+				+ "normalize=True, "
+				+ "batch_size=8, "
+				+ "flow3D_smooth=0, "
+				+ "flow_threshold=" + flowThreshold + ", "
+				+ "cellprob_threshold=" + cellprobThreshold + ")" + "\n";
 	}
 
 	@Override
@@ -41,31 +72,6 @@ public class Cellpose3 extends Segmentation3D
 				+ "  - pytorch\n"
 				+ "  - pytorch-cuda\n"
 				+ "  - numpy\n";
-	}
-
-	@Override
-	String generateScript()
-	{
-		return "import numpy as np" + "\n"
-				+ "from cellpose import models" + "\n"
-				+ "import appose" + "\n"
-				+ "\n"
-				+ "task.update(message=\"Imports completed\")" + "\n"
-				+ "image_ndarray = image.ndarray()" + "\n"
-				+ "task.update(message=\"Image converted, Image shape: \" + \",\".join(str(dim) for dim in image.shape))" + "\n"
-				+ getLoadModelCommand()
-				+ "task.update(message=\"Model loaded\")" + "\n"
-				+ "\n"
-				+ getEvaluateModelCommand()
-				+ "\n"
-				+ "task.update(message=\"Segmentation completed, Segmentation shape: \" + \",\".join(str(dim) for dim in segmentation.shape))"
-				+ "\n"
-				+ "\n"
-				+ "shared = appose.NDArray(image.dtype, image.shape)" + "\n"
-				+ "task.update(message=\"NDArray created\")" + "\n"
-				+ "shared.ndarray()[:] = segmentation" + "\n"
-				+ "task.update(message=\"NDArray filled\")" + "\n"
-				+ "task.outputs['label_image'] = shared" + "\n";
 	}
 
 	public enum MODEL_TYPE
@@ -95,7 +101,6 @@ public class Cellpose3 extends Segmentation3D
 		LC4( "LC4", false );
 
 		private final String modelName;
-
 		private final boolean hasSizeModel;
 
 		MODEL_TYPE( final String modelName, final boolean hasSizeModel )
@@ -131,80 +136,5 @@ public class Cellpose3 extends Segmentation3D
 			}
 			throw new IllegalArgumentException( "No enum constant for model name: " + modelName );
 		}
-	}
-
-	public double getAnisotropy()
-	{
-		return anisotropy;
-	}
-
-	public void setAnisotropy( final double anisotropy )
-	{
-		this.anisotropy = anisotropy;
-	}
-
-	public double getCellprobThreshold()
-	{
-		return cellprobThreshold;
-	}
-
-	public void setCellprobThreshold( final double cellprobThreshold )
-	{
-		this.cellprobThreshold = cellprobThreshold;
-	}
-
-	public double getFlowThreshold()
-	{
-		return flowThreshold;
-	}
-
-	public void setFlowThreshold( final double flowThreshold )
-	{
-		this.flowThreshold = flowThreshold;
-	}
-
-	public boolean is3D()
-	{
-		return is3D;
-	}
-
-	public void set3D( final boolean is3D )
-	{
-		this.is3D = is3D;
-	}
-
-	private String is3DParam()
-	{
-		return is3D ? "True" : "False";
-	}
-
-	private String anisotropyParam()
-	{
-		return is3D() ? String.valueOf( anisotropy ) : "1.0";
-	}
-
-	private String getLoadModelCommand()
-	{
-		if ( modelType.hasSizeModel() )
-			return "model = models.Cellpose(model_type=\"" + modelType.getModelName() + "\", gpu=True)" + "\n";
-		else
-			return "model = models.CellposeModel(model_type=\"" + modelType.getModelName() + "\", gpu=True)" + "\n";
-	}
-
-	private String getEvaluateModelCommand()
-	{
-		String diams = modelType.hasSizeModel() ? ", diams" : "";
-		return "segmentation, flows, styles" + diams + " = model.eval("
-				+ "image_ndarray, "
-				+ "diameter=None, "
-				+ "channels=[0, 0], "
-				+ "do_3D=" + is3DParam() + ", "
-				+ "anisotropy=" + anisotropyParam() + ", "
-				+ "z_axis=0, "
-				+ "normalize=True, "
-				+ "batch_size=8, "
-				+ "flow3D_smooth=0, "
-				+ "flow_threshold=" + flowThreshold + ", "
-				+ "cellprob_threshold=" + cellprobThreshold + ")" + "\n";
 	}
 }
